pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "ecip-api-gateway"
version = "0.1.0"
description = ""
authors = []

[tool.poetry.dependencies]
PyYAML = "^6.0"
asgi-correlation-id = "^3.0.1"
email-validator = "^1.2.1"
fastapi = "^0.83.0"
fastapi-utils = "^0.2.1"
fastapi-versioning = "^0.10.0"
gunicorn = "^20.1.0"
httpx = "^0.23.0"
loguru = "^0.6.0"
orjson = "^3.8.0"
phonenumbers = "^8.12.55"
pydantic-factories = "^1.17.0"
pytest = "^7.1.3"
pytest-anyio = "^0.0.0"
pytest-mock = "^3.10.0"
python = "^3.11.5"
python-dateutil = "^2.8.2"
python-json-logger = "^2.0.4"
uvicorn = "^0.18.3"
python-multipart = "^0.0.5"
pydantic = "^1.10"
typer = "^0.7.0"

[tool.poetry.group.dev.dependencies]
autoflake = "^1.5.3"
black = "^22.8.0"
devtools = "^0.9.0"
isort = "^5.10.1"
pre-commit = "^2.20.0"
pylint = "^2.15.2"
pylint-quotes = "^0.2.3"
remove-print-statements = "^0.5.0"
sqlalchemy-stubs = "^0.4"
sqlalchemy2-stubs = "^0.0.2-alpha.27"
yapf = "^0.3.1"

[tool.black]
line-length = 100
skip-string-normalization = true

[tool.isort]
profile = "black"
line_length = 100

[tool.pylint.master]
init-hook='import sys; sys.path.append(".")'

[tool.pylint.messages_control]
max-line-length = 100
extension-pkg-whitelist = "pydantic"
disable = [
    "R",
    "arguments-differ",
    "broad-except",
    "fixme",
    "invalid-name",
    "logging-fstring-interpolation",
    "missing-class-docstring",
    "missing-function-docstring",
    "missing-module-docstring",
    "no-self-argument",
    "pointless-statement",
    "protected-access",
    "raise-missing-from",
    "singleton-comparison",
    "super-init-not-called",
    "too-many-lines",
    "unidiomatic-typecheck",
    "unnecessary-lambda",
    "unused-argument",
    "unused-import",
    "line-too-long",
]
